Closes #3 se crearon paquetes y se refactorizaron clases
