--- app/src/main/java/com/g5/tdp2/cashmaps/BankGateway.java
@@ -1,7 +1,0 @@
-package com.g5.tdp2.cashmaps;
-
-import java.util.List;
-
-public interface BankGateway {
-    List<String> getBanks();
-}
