+++ app/src/main/java/com/g5/tdp2/cashmaps/gateway/AtmGatewayException.java
@@ -1,4 +1,4 @@
-package com.g5.tdp2.cashmaps;
+package com.g5.tdp2.cashmaps.gateway;
 
 public class AtmGatewayException extends RuntimeException {
     public AtmGatewayException(String message, Throwable cause) {
