+++ app/src/main/java/com/g5/tdp2/cashmaps/gateway/BankGateway.java
@@ -1,0 +1,15 @@
+package com.g5.tdp2.cashmaps.gateway;
+
+import java.util.List;
+
+/**
+ * Obtiene el catalogo de bancos disponibles
+ */
+public interface BankGateway {
+    /**
+     * Obtiene el catalogo de bancos disponibles
+     *
+     * @return catalogo de bancos disponibles
+     */
+    List<String> getBanks();
+}
